requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-auth==2.29.0
google-api-python-client==2.128.0
